FORMAT: Organize folder structure with namespace
