+++ tsconfig.json
@@ -2,7 +2,7 @@
 	"compilerOptions": {
 		/* Basic Options */
 		"target": "es6" /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017','ES2018' or 'ESNEXT'. */,
-		"module": "commonjs" /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */,
+		"module": "amd" /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */,
 		"lib": [
 			"dom",
 			"es6",
@@ -15,7 +15,7 @@
 		// "declaration": true,                   /* Generates corresponding '.d.ts' file. */
 		// "declarationMap": true,                /* Generates a sourcemap for each corresponding '.d.ts' file. */
 		"sourceMap": true /* Generates corresponding '.map' file. */,
-		// "outFile": "./",                       /* Concatenate and emit output to single file. */
+		"outFile": "./dist/bundle.js" /* Concatenate and emit output to single file. */,
 		"outDir": "./dist" /* Redirect output structure to the directory. */,
 		"rootDir": "./src" /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
 		// "composite": true,                     /* Enable project compilation */
